requirements.txt
lxml==4.8.0
Pillow==9.1.0
python-pptx==0.6.21
XlsxWriter==3.0.3
